Rewrite main loop
I haven’t benchmarked it, but from a clarity perspective, I’m not
at all sure it’s better. I think what I had in mind when I wanted
to try is something that can be built with a traditional for loop
and a function that returns the first UTF-8 encoding from a string
and its width. In Rust, the traditional for loop is omitted because
we’re not smart enough to use it, and that function is embedded in
an iterator. The Rust way may be to do the loop with an iterator.
I haven’t tried that yet, so I can’t comment on the clarity. Pretty
sure it’ll be higher ceremony, though.